requirements.txt
google-genai
pydantic
python-dotenv
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6
websockets>=12.0
rich>=13.0.0
openpyxl>=3.1.0
